+++ Plugins/CavrnusConnector/Source/CavrnusConnector/Private/Avatar/CavrnusSyncAvatarTransform.cpp
@@ -1,3 +1,0 @@
-// Copyright (c), Cavrnus. All rights reserved.
-
-#include "Avatar/CavrnusSyncAvatarTransform.h"