+++ Plugins/CavrnusConnector/Source/CavrnusConnector/Private/CavrnusPropertiesContainer.cpp
@@ -1,4 +1,4 @@
-// Copyright(c) Cavrnus. All rights reserved.
+// Copyright (c) 2024 Cavrnus. All rights reserved.
 
 #include "CavrnusPropertiesContainer.h"
 #include "CavrnusSpatialConnectorSubSystem.h"