+++ Plugins/CavrnusConnector/Source/CavrnusConnector/Private/RelayModel/CavrnusRelayModel.cpp
@@ -1,4 +1,4 @@
-﻿// Copyright(c) Cavrnus. All rights reserved.
+﻿// Copyright (c) 2024 Cavrnus. All rights reserved.
 
 #include "RelayModel/CavrnusRelayModel.h"
 #include "CavrnusConnectorSettings.h"
@@ -7,6 +7,7 @@
 #include "../Interop/CavrnusInteropLayer.h"
 #include "RelayModel/RelayCallbackModel.h"
 #include "RelayModel/DataState.h"
+#include "RelayModel/CavrnusDataCache.h"
 #include <HAL/PlatformTime.h>
 #include "CoreMinimal.h"
 #include "Types/CavrnusSpawnedObject.h"
@@ -33,6 +34,7 @@ namespace Cavrnus
 
 		dataState = new DataState();
 		callbackModel = new RelayCallbackModel(this);
+		dataCache = new CavrnusDataCache();
 	}
 
 	CavrnusRelayModel::~CavrnusRelayModel()
@@ -132,6 +134,11 @@ namespace Cavrnus
 		return dataState;
 	}
 
+	CavrnusDataCache* CavrnusRelayModel::GetDataCache()
+	{
+		return dataCache;
+	}
+
 
 	void CavrnusRelayModel::SendMessage(const ServerData::RelayClientMessage& msg)
 	{
@@ -172,6 +179,9 @@ namespace Cavrnus
 
 		switch (msg.Msg_case())
 		{
+		case ServerData::RelayRemoteMessage::kRelayDataCache:
+			HandleInitialDataCache(msg.relaydatacache());
+			break;
 		case ServerData::RelayRemoteMessage::kPropertyValueStatus:
 			HandleServerPropertyUpdate(msg.propertyvaluestatus());
 			break;
@@ -180,6 +190,9 @@ namespace Cavrnus
 			break;
 		case ServerData::RelayRemoteMessage::kAuthenticateGuestResp:
 			callbackModel->HandleServerCallback(msg.authenticateguestresp().reqid(), msg);
+			break;
+		case ServerData::RelayRemoteMessage::kAuthenticateTokenResp:
+			callbackModel->HandleServerCallback(msg.authenticatetokenresp().reqid(), msg);
 			break;
 		case ServerData::RelayRemoteMessage::kJoinSpaceFromIdResp:
 			callbackModel->HandleServerCallback(msg.joinspacefromidresp().reqid(), msg);
@@ -195,6 +208,9 @@ namespace Cavrnus
 			break;
 		case ServerData::RelayRemoteMessage::kContentDestinationFolderResp:
 			callbackModel->HandleServerCallback(msg.contentdestinationfolderresp().reqid(), msg);
+			break;
+		case ServerData::RelayRemoteMessage::kGetSpaceInfoResp:
+			callbackModel->HandleServerCallback(msg.getspaceinforesp().reqid(), msg);
 			break;
 		case ServerData::RelayRemoteMessage::kMessage:
 			HandleLogging(msg.message());
@@ -284,6 +300,20 @@ namespace Cavrnus
 		default:
 			break;
 		}
+	}
+
+	void CavrnusRelayModel::HandleInitialDataCache(const ServerData::RelayDataCache& cache)
+	{
+		TArray<FString> keys;
+		TArray<FString> vals;
+		for (int i = 0; i < cache.stringkeys_size(); i++)
+		{
+			keys.Add(UTF8_TO_TCHAR(cache.stringkeys()[i].c_str()));
+			vals.Add(UTF8_TO_TCHAR(cache.stringvalues()[i].c_str()));
+		}
+		GetDataCache()->Setup(keys, vals);
+
+		callbackModel->HandleDataCache(cache);
 	}
 
 	void CavrnusRelayModel::HandleSpaceUserAdded(ServerData::CavrnusSpaceConnection spaceConn, ServerData::CavrnusUser user)